Job config (#175)
* Extract interface forwarding to separate function

Extract the forwarding of dataset identifiers to the interface into a
new function to allow it to be more self-contained.

* Insert identifiers upon forward failures

When forwarding to the interface, check whether the operation fails or
not. If it does, attempt to insert the identifier into a MongoDB
collection for persistent storage.

* Add a struct for interface jobs

Add a struct to manage the structure of interface jobs.

* Rename collection name to be simpler

* Eof check (#168)

* split protocol into own mod

* created auth_end

* fixed EOF issue

* created ClientMessage and InterfaceMessage

* Extract interface forwarding to separate function

Extract the forwarding of dataset identifiers to the interface into a
new function to allow it to be more self-contained.

* Insert identifiers upon forward failures

When forwarding to the interface, check whether the operation fails or
not. If it does, attempt to insert the identifier into a MongoDB
collection for persistent storage.

* Add a struct for interface jobs

Add a struct to manage the structure of interface jobs.

* created ClientMessage and InterfaceMessage

* Integrated InterfaceMessage

changing so that all messages over interface should be interface
messages. Currently only a Config message but has capacity for
more in the future.

* writing config to node

* Add comments

* added inferred columns to configs

* sending proper columns types to client

* written frontend job config dropdown

* Upgrade the DCL to use Tokio 1.0

* Use a GitHub version of MongoDB

* Update all Tokio versions to 1.0

* Remove DCL specific messages

Remove the DCL specific messages as the messages module is updated to
Tokio 1.0 now.

* Update interface to use Tokio 1.0

Update the interface to be asynchronous, using Tokio 1.0 for its runtime
inline with most other components.

* Change the interface to use channels

Change the interface layer to use Tokio's channels instead of a queue,
as the blocking Mutex causes task execution to stop.

* Prefer tokio::time::sleep over thread::sleep

Asynchronous processes will be blocked if using thread::sleep as they
run on a singular thread. Instead, tokio::time::sleep does the same
thing asynchronously and thus allows other tasks to make progress.

* Clean up logging output slightly

* Api tokio into job_config (#178)

* converted most to actix

* fixed common test stuff

* tests don't work and it panics

* added actix and changed rt version

* added arcs

* solved filter issue

* error implementing factory thing

* add await

* converted code base to use other route reg

* getting tests working

* projects tests all pass

* tests all pass

* finalising api_tokio changes

* Upgrade Tokio 0.3 to 1.0

* Remove tokio-util as it isn't needed

* Upgrade all dependencies to be Tokio 1.0

Upgrade all the `actix-*` dependencies to be compatible with Tokio 1.0
by using beta versions, and use the Git dependency for MongoDB for now.

* Change DodonaError::name to return &'static str

Change the `DodonaError::name` to return a `&'static str` instead of a
String as allocations are not needed.

* Refactor the length prefixing into a trait (#180)

Length prefixing for messages previously shared an implementation
between both message types. This left it vulnerable to one being changed
without the update being propagated.

Instead, this adds a `ReadLengthPrefix` trait for reading a length
prefixed message from a stream and a `WriteLengthPrefix` trait for
converting a type into a vector of bytes prefixed with its length.

In some cases, we want to length prefix a message but not contain the
type information, such as when acting as a JSON proxy for the API
server. For this, the `RawMessage` type acts as a wrapper, only
serializing its contents.

* made building a cluster more resiliant and mdb tokio 1.0 (#181)

* made building a cluster more resiliant

* cleaning up

* Fix a lot of `clippy` warnings

* Address review comments

* Fix merge issues

+++ backend/api-server/tests/common.rs
@@ -1,7 +1,9 @@
 use std::str::FromStr;
+use std::sync::Arc;
 
 use mongodb::bson::{self, document::Document, oid::ObjectId};
 
+use api_server::AppState;
 use config::Environment;
 use models::projects::Project;
 use models::users::User;
@@ -21,7 +23,7 @@ pub static DELETABLE_PROJECT_ID: &str = "5fb2b4049d524e99ac7f1c41";
 pub static EDITABLE_PROJECT_ID: &str = "5fb2c4e4b4b7becc1e81e278";
 
 /// Allows for the setup of the database prior to testing.
-static INIT: std::sync::Once = std::sync::Once::new();
+static MUTEX: tokio::sync::Mutex<bool> = tokio::sync::Mutex::const_new(true);
 
 /// Defines the initialisation function for the tests.
 ///
@@ -31,38 +33,60 @@ static INIT: std::sync::Once = std::sync::Once::new();
 ///
 /// As the database can be initialised before running, this allows tests to be run in any order
 /// provided they don't require the result of a previous test.
-pub fn initialise() {
-    INIT.call_once(|| {
-        async_std::task::block_on(async {
-            // Setup the environment variables
-            let config = config::ConfigFile::from_filesystem();
-            let resolved = config.resolve(Environment::Testing);
-            resolved.populate_environment();
+pub async fn initialise() -> AppState {
+    // Aquire the mutex
+    let mut lock = MUTEX.lock().await;
 
-            // Connect to the database
-            let conn_str = std::env::var("CONN_STR").expect("CONN_STR must be set");
+    // Check whether this is the first time being run
+    if *lock {
+        let config = config::ConfigFile::from_filesystem();
+        let resolved = config.resolve(Environment::Testing);
+        resolved.populate_environment();
 
-            // Ensure that we aren't using the Atlas instance
-            assert!(
-                !conn_str.starts_with("mongodb+srv"),
-                "Please setup a local MongoDB instance for running the tests"
-            );
+        // Connect to the database
+        let conn_str = std::env::var("CONN_STR").expect("CONN_STR must be set");
 
-            let client = mongodb::Client::with_uri_str(&conn_str).await.unwrap();
-            let database = client.database("sybl");
-            let collection_names = database.list_collection_names(None).await.unwrap();
+        // Ensure that we aren't using the Atlas instance
+        assert!(
+            !conn_str.starts_with("mongodb+srv"),
+            "Please setup a local MongoDB instance for running the tests"
+        );
+        let client = mongodb::Client::with_uri_str(&conn_str).await.unwrap();
+        let database = client.database("sybl");
+        let collection_names = database.list_collection_names(None).await.unwrap();
 
-            // Delete all records currently in the database
-            for name in collection_names {
-                let collection = database.collection(&name);
-                collection.delete_many(Document::new(), None).await.unwrap();
-            }
+        // Delete all records currently in the database
+        for name in collection_names {
+            let collection = database.collection(&name);
+            collection.delete_many(Document::new(), None).await.unwrap();
+        }
 
-            // Insert some test data
-            insert_test_users(&database).await;
-            insert_test_projects(&database).await;
-        });
-    });
+        // Insert some test data
+        insert_test_users(&database).await;
+        insert_test_projects(&database).await;
+
+        // Update the lock
+        *lock = false;
+    }
+
+    build_app_state().await
+}
+
+pub async fn build_app_state() -> AppState {
+    let conn_str = std::env::var("CONN_STR").expect("CONN_STR must be set");
+    let pepper = std::env::var("PEPPER").expect("PEPPER must be set");
+    let pbkdf2_iterations =
+        std::env::var("PBKDF2_ITERATIONS").expect("PBKDF2_ITERATIONS must be set");
+
+    let client = mongodb::Client::with_uri_str(&conn_str).await.unwrap();
+
+    AppState {
+        client: Arc::new(client.clone()),
+        db_name: Arc::new(String::from("sybl")),
+        pepper: Arc::new(pepper.clone()),
+        pbkdf2_iterations: u32::from_str(&pbkdf2_iterations)
+            .expect("PBKDF2_ITERATIONS must be parseable as an integer"),
+    }
 }
 
 fn create_user_with_id(
@@ -156,26 +180,4 @@ async fn insert_test_projects(database: &mongodb::Database) {
     projects.insert_one(overwritten_data, None).await.unwrap();
     projects.insert_one(deletable, None).await.unwrap();
     projects.insert_one(editable, None).await.unwrap();
-}
-
-pub fn build_json_request(url: &str, body: &str) -> tide::http::Request {
-    let full_url = format!("localhost:{}", url);
-    let url = tide::http::Url::parse(&full_url).unwrap();
-    let mut req = tide::http::Request::new(tide::http::Method::Post, url);
-
-    req.set_body(body);
-    req.set_content_type(tide::http::mime::JSON);
-
-    req
-}
-
-pub fn build_json_put_request(url: &str, body: &str) -> tide::http::Request {
-    let full_url = format!("localhost:{}", url);
-    let url = tide::http::Url::parse(&full_url).unwrap();
-    let mut req = tide::http::Request::new(tide::http::Method::Put, url);
-
-    req.set_body(body);
-    req.set_content_type(tide::http::mime::JSON);
-
-    req
 }
